requirements.txt
#LangChin Core
langchain
langchain-core

#OpenAI Integration
langchain-openai
openai

#Anthropic Integration
langchain-anthropic

#google gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Environment variable Management
python-dotenv

#Machine Learning Utilities
numpy
scikit-learn

